requirements.txt
pycryptodome==3.18.0
Requests==2.31.0
